+++ packages/launchpad/cypress/e2e/scaffold-project.cy.ts
@@ -37,6 +37,7 @@ function scaffoldAndOpenE2EProject (opts: {
   }
 
   cy.visitLaunchpad()
+  cy.skipWelcome()
 
   cy.contains('Welcome to Cypress!').should('be.visible')
   cy.contains('[data-cy-testingtype="e2e"]', 'Not Configured')
@@ -71,6 +72,7 @@ function scaffoldAndOpenCTProject (opts: {
   }
 
   cy.visitLaunchpad()
+  cy.skipWelcome()
 
   cy.contains('Welcome to Cypress!').should('be.visible')
   cy.contains('[data-cy-testingtype="e2e"]', 'Not Configured')
@@ -174,6 +176,7 @@ describe('scaffolding new projects', { defaultCommandTimeout: 7000 }, () => {
     })
 
     cy.visitLaunchpad()
+    cy.skipWelcome()
     cy.contains('button', cy.i18n.testingType.e2e.name).click()
     cy.contains('button', cy.i18n.setupPage.step.continue).click()
     cy.contains('h1', cy.i18n.setupPage.testingCard.chooseABrowser).should('be.visible')